# labstack-echo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A sample server that uses the Go web framework Echo.",
        "title": "Echo example",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/fi3te/hello-go"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/users/:id": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Returns a specific user",
                "operationId": "getUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of the user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.User": {
            "type": "object",
            "properties": {
                "forename": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "surename": {
                    "type": "string"
                }
            }
        }
    }
}